requirements.txt
browser-use==0.1.48
langchain-openai
python-dotenv
playwright
